LG-11116 Refactor Report Building
[skip changelog]
